try to solve bug

+++ src/rsfml/graphics/render_window.rs
@@ -391,7 +391,7 @@ impl RenderWindow {
                     0 => false,
                     _ => true
                 };
-                let k : keyboard::Key = unsafe { cast::transmute(self.event.p1 as i8) };
+                let k : keyboard::Key = unsafe { cast::transmute(self.event.p1 as i64) };
                 event::KeyPressed{ 
                     code : k, 
                     alt : al, 
@@ -417,7 +417,7 @@ impl RenderWindow {
                     0 => false,
                     _ => true
                 };
-                let k : keyboard::Key = unsafe { cast::transmute(self.event.p1 as i8) };
+                let k : keyboard::Key = unsafe { cast::transmute(self.event.p1 as i64) };
                 event::KeyReleased {
                     code : k, 
                     alt : al,
